+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ features.
 
 - Added :func:`brownie.functional.bind_arguments`.
 - Allow using :func:`brownie.itools.unique` with non-hashable items.
+- Added :mod:`pickle` support to :class:`brownie.datastructures.LazyList`.
 
 0.4.1
 -----